+++ src/main/java/org/jufe/anmeldetool/service/message/MessageStore.java
@@ -1,0 +1,13 @@
+package org.jufe.anmeldetool.service.message;
+
+import org.springframework.ui.Model;
+
+import java.util.HashMap;
+
+public class MessageStore extends HashMap<String, String> {
+
+    public void addToModel(Model model) {
+        model.addAllAttributes(this);
+    }
+
+}
